In delPaths also delete the possible reverse path
Take e.g. 2 nodes. A and B. Because of the implementation, both nodes
will report each other in their routing updates so A will have a route
to B and B will have route to A. In delPahts, only one route was
removed when a node was removed. This fix will also remove the possible
reverse route.